+++ src/objects/game/obj_background.ts
@@ -1,0 +1,11 @@
+import { k } from "../../main";
+
+export const backgroundObj = (color: string) => {
+    const background = k.make([
+        k.layer("background"),
+        k.rect(k.width(), k.height()),
+        k.color(k.Color.fromHex("#ee8fcb"))
+    ]);
+
+    return background;
+}