+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# Alpha 1.2.0 - 7/1/2024
+## Mechanics
+- New note type: "Slider"
+- Added support for change the notes velocity (#SCROLL) and change the measure (#MEASURE)
+
+## Graphics
+- New slider sprite: "Green Apple"
+- New song in main menu desing
+
 # Alpha 1.1.0 - 12/11/2023
 ## Mechanics
 - Combo